feat: refactor footer component and styles into molecules
