+++ Cardoon/src/components/pages/GamePage/GamePageStyle.scss
@@ -32,70 +32,6 @@
       right: 1rem;
     }
   }
-
-  &__footer {
-    position: fixed;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    display: flex;
-    gap: 1rem;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0.5rem 2rem;
-    background-color: #f8f9fa;
-    border-top: 1px solid #dee2e6;
-
-    &__element {
-      display: flex;
-      align-items: center;
-      font-size: 1.2rem;
-      color: #343a40;
-    }
-
-    .GamePage__icon {
-      width: 24px;
-      height: 24px;
-      margin-right: 0.25rem;
-    }
-
-    &__flashmode {
-      margin: 0 !important;
-    }
-
-    &__progress-bar {
-      width: 200px;
-      background-color: #e9ecef;
-      border-radius: 0.25rem;
-      overflow: hidden;
-      position: relative;
-
-      &__title {
-        font-size: 1rem;
-        color: #343a40;
-        margin-right: 0.5rem;
-
-        @media (max-width: 768px) {
-          display: none; // Hide title on mobile
-        }
-      }
-      &__fill {
-        padding: 0.75rem 0;
-        background-color: #007bff;
-        transition: width 0.5s ease-in-out; /* Smooth transition for progress */
-      }
-
-      &__text {
-        padding: 0.5rem;
-        font-size: 0.875rem;
-        color: #374151;
-        position: absolute;
-        top: 50%;
-        left: 50%;
-        transform: translate(-50%, -50%);
-      }
-    }
-  }
 }
 .floating-coin {
   width: 24px;
