+++ Cardoon/src/components/pages/UpgradePage/UpgradePage.tsx
@@ -3,7 +3,7 @@ import { UserContext } from "../../../context/UserContext";
 import { useFetch, RESOURCES } from "../../../hooks/server";
 import { Item as ItemType, UserItem } from "../../../types/common";
 import Item from "../../molecules/Item/Item";
-import { GameFooter } from "../../Footer/Footer";
+import { GameFooter } from "../../molecules/Footer/Footer";
 
 export default () => {
   const { user, hasItem, refresh } = useContext(UserContext);
